Remove all padding (no-operations)
The purpose of all no-operations was not alignment, but rather padding to
allow insertion of more instructions anywhere. Now that the assembling is
automated and no longer done by hand, this padding is obsolete and its
removal resulted in around 230 bytes of more space. This will aid in
accomplishing the implementation of planned features and improvements.