Add console command to display active patches
By typing 'ninja' (without quotes) into the game console, the version, as
well as all active patches (sorted), if any, are displayed.
Additionally, an issue from the previous commit is fixed, that caused
abnormal termination when no Ninja patch was active.
The core was moved yet again to a larger address space.