Add G1/G2 postfix to SRC to unify patches
The same patch now works for both Gothic and Gothic 2 NotR as Ninja looks
for a game specific postfix in the SRC and animation file names. This
will, however, increase the amount of SRC files in the patch and renders
all previous Ninja patches incompatible with the comming Ninja version.

Fix #1