Give warning for outdated patches
Since Ninja used to be part of the VDF of patches and is injected by the
SystemPack after the new implementation, it overwrites the new version.
Outdated patches are hereby detected and the game is terminated with a
corresponding error message.

The executive and function blocks of the core are switched to have the
crucial code for this detection itself not overwritten.