Reduce core and fix hook in Hlp_GetNpc
The core was overlapping a function (that is only called with zSpy level
> 7). Now it is reduced by about 300 bytes to remain within acceptable
address bounds. Mostly, this was achieved by removing output messages
and utilizing pusha/popa instead of pushing individual registers.

Additionally, a severe bug in the Hlp_GetNpc hook is now fixed, that had
caused a crash when the helper symbol yInstance_Help had never been set
before and Hlp_GetNpc was called with 0.