Exchange DLL wrapper
The orignal BugslayerUtil.DLL is expected in the same directory with the
name BugslayerUtilG.DLL. Ninja loads it when first needed. This usually
happens from SetCrashHandlerFilter. The DLL is never detached
afterwards.

If the original file cannot be found, simplistic versions of the
exported functions return appropriate values for the game to close
properly. Neither the registers, nor a stack trace will be printed, but
messages indicating that the file is missing instead.

Fix #14