Attempt to recover from invalid spell items
Gothic encounters an unrecoverable error when unarchiving oCNpcInventory
and facing an equipped spell item (scroll or rune) that no longer exists
in the game scripts. This is due to an extra archive-value that will not
be skipped when continuing to the next item. While this error is native
to Gothic, it is more likely to happen if a patch introduces a spell
item, that item is equipped and the patch is removed. While corrupt
binary archives are unrecoverable, and the game will crash with
certainty, this commit introduces an attempt to recover, by advancing
the buffer beyond a potential extra archive-value and tries to read the
item object again. Note that this will guarantee to solve the above
issue, but that issue only.