Recover last invalid spell item
This commit complements the previous one to attempt to fix an
unrecoverable error of equipped invalid spell items. What the previous
commit lacked was the recovery if the last equipped item (in order of
inventory sorting) is an invalid spell item. Then the corrupted
reading position was carried into the next unarchiving processes. Here
this is attempted to fix by checking the following values from
unarchiving.